- Use a 64-bit int for the baudrate within the ifmedia_baudrate struct and return a 64-bit int for ifmedia_baudrate(). - Fix consumers of ifmedia_baudrate() to deal with the change.
ok claudio@ dlg@

+++ sys/net/if_media.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: if_media.h,v 1.23 2006/12/29 20:06:31 kettenis Exp $	*/
+/*	$OpenBSD: if_media.h,v 1.24 2007/05/29 22:11:57 henning Exp $	*/
 /*	$NetBSD: if_media.h,v 1.22 2000/02/17 21:53:16 sommerfeld Exp $	*/
 
 /*-
@@ -145,7 +145,7 @@ struct	ifmedia_entry *ifmedia_match(struct ifmedia *, u_int, u_int);
 void	ifmedia_delete_instance(struct ifmedia *, u_int);
 
 /* Compute baudrate for a given media. */
-int	ifmedia_baudrate(int);
+u_int64_t	ifmedia_baudrate(int);
 #endif /*_KERNEL */
 
 /*
@@ -514,7 +514,7 @@ struct ifmedia_description {
  */
 struct ifmedia_baudrate {
 	int	ifmb_word;		/* media word */
-	int	ifmb_baudrate;		/* corresponding baudrate */
+	u_int64_t	ifmb_baudrate;		/* corresponding baudrate */
 };
 
 #define	IFM_BAUDRATE_DESCRIPTIONS {					\
